% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_basemap.R
\name{add_basemap}
\alias{add_basemap}
\title{Add basemaps to leaflet maps with light or dark theme}
\usage{
add_basemap(map, dark = FALSE, layers = FALSE)
}
\arguments{
\item{dark}{TRUE/FALSE - Use a dark theme? Default is FALSE.}

\item{layers}{TRUE/FALSE - Add multiple basemap layers? Default is FALSE.}
}
\description{
Add basemaps to leaflet maps with light or dark theme
}
\examples{
library(leaflet)
library(basemap)
\dontrun{
df <- ozone_aqi 

# Light background    
leaflet(df) \%>\%
  addCircleMarkers(fillColor   = ~aqi_color, 
                   color       = 'gray',
                   fillOpacity = 0.8) \%>\%
  add_basemap()

# Dark background    
leaflet(df) \%>\%
  addCircleMarkers(fillColor = ~aqi_color, 
                   color     = 'gray',
                   fillOpacity = 0.8) \%>\%
  add_basemap(dark = TRUE)
  
  # Add multiple basemap layers    
  leaflet(df) \%>\%
  addCircleMarkers(fillColor = ~aqi_color, 
                   color     = 'gray',
                   fillOpacity = 0.8) \%>\%
  add_basemap(layers = TRUE)
}

}
