% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_basemaps.R
\name{add_basemaps}
\alias{add_basemaps}
\title{Easy add basemaps to leaflet that the user can switch On/Off}
\usage{
add_basemaps(type = "light", controls = TRUE)
}
\arguments{
\item{type}{"light" or "dark"}

\item{controls}{TRUE/FALSE - Let the user switch base maps}
}
\description{
Easy add basemaps to leaflet that the user can switch On/Off
}
\examples{
library(leaflet)
library(basemap)
\dontrun{

leaflet(my_map) \%>\%
  add_basemaps()

}

}
